@@ -4,21 +4,19 @@
   "private": true,
   "dependencies": {
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^12.0.0",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.11.41",
     "@types/react": "^18.0.14",
     "@types/react-dom": "^18.0.5",
-    "onsenui": "^2.12.1",
     "postcss-flexbugs-fixes": "^5.0.2",
     "postcss-normalize": "^10.0.1",
     "postcss-preset-env": "^7.7.2",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-chatbot-kit": "^2.0.1",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-markdown": "^8.0.3",
-    "react-onsenui": "^1.12.0",
     "react-scripts": "5.0.1",
     "remark-gfm": "^3.0.1",
     "typescript": "^4.7.4",
